scoring-service/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
redis==5.1.1
kafka-python==2.0.2
xgboost==2.1.1
scikit-learn==1.5.2
numpy==2.1.0
joblib==1.4.2
